Updated helper to model, error handling and camel case
